refactor: 로또 생성 방법에 전략 패턴 적용
