+++ src/main/java/domain/lottogenerator/LottoGenerator.java
@@ -1,0 +1,8 @@
+package domain.lottogenerator;
+
+import domain.LottoNumber;
+import java.util.List;
+
+public interface LottoGenerator {
+    List<LottoNumber> generate();
+}
