+++ src/test/java/domain/LottosTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.lottogenerator.RandomLottoGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +11,7 @@ class LottosTest {
     @Test
     void lottosCreationTest() {
         int quantity = 7;
-        Lottos lottos = Lottos.ofSize(quantity);
+        Lottos lottos = Lottos.ofSize(quantity, new RandomLottoGenerator());
 
         // TODO: 사이즈 체크하기
         assertThat(lottos).isNotNull()
